% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsim.R
\name{mlsim}
\alias{mlsim}
\title{Library and function for testing ML classification algorithms trained on simulated data}
\usage{
mlsim(lwr = -5, upr = 5, N = 500)
}
\arguments{
\item{lwr}{lower bound for numerical data}

\item{upr}{upper bound for numerical data}

\item{N}{sample size}
}
\description{
Generate random data and apply different ML algorithms
}
\details{
In order to see the list of all private functions used in this library type mlsim::: and then hit the Tab key
}
\author{
Jacek Kowalski, \email{@australianunity.com.au}
}
